LayoutTests/platform/ios-simulator/css3/shapes/shape-outside/values/shape-outside-computed-shape-000-expected.txt


PASS em units 
FAIL ex units assert_equals: expected "polygon(71.516px 71.516px)" but got "polygon(71.52px 71.52px)"(stack: assert@file:///Volumes/Data/Development/OSX/webkit/OpenSource/LayoutTests/resources/testharness.js:2168:73
assert_equals@file:///Volumes/Data/Development/OSX/webkit/OpenSource/LayoutTests/resources/testharness.js:705:15
testComputedStyle@file:///Volumes/Data/Development/OSX/webkit/OpenSource/LayoutTests/css3/shapes/shape-outside/values/support/parsing-utils.js:24:22
testUnit@file:///Volumes/Data/Development/OSX/webkit/OpenSource/LayoutTests/css3/shapes/shape-outside/values/shape-outside-computed-shape-000.html:42:47
file:///Volumes/Data/Development/OSX/webkit/OpenSource/LayoutTests/resources/testharness.js:477:40
step@file:///Volumes/Data/Development/OSX/webkit/OpenSource/LayoutTests/resources/testharness.js:1186:30
test@file:///Volumes/Data/Development/OSX/webkit/OpenSource/LayoutTests/resources/testharness.js:429:22
file:///Volumes/Data/Development/OSX/webkit/OpenSource/LayoutTests/resources/testharness.js:475:25
forEach@file:///Volumes/Data/Development/OSX/webkit/OpenSource/LayoutTests/resources/testharness.js:2231:30
generate_tests@file:///Volumes/Data/Development/OSX/webkit/OpenSource/LayoutTests/resources/testharness.js:472:16
global code@file:///Volumes/Data/Development/OSX/webkit/OpenSource/LayoutTests/css3/shapes/shape-outside/values/shape-outside-computed-shape-000.html:49:27)
PASS ch units 
PASS rem units 
